Merge branch 'devel-sync' into 'main'

See merge request msu-libraries/devops/catalog!47

+++ vufind/startup.sh
@@ -1,19 +1,21 @@
 #!/bin/bash
 
-# Create symlinks to the shared storage
+# Create symlinks to the shared storage for non-production environments
 # Populating the shared storage if empty
-mkdir -p /mnt/shared/local/${STACK_NAME}
-if [ ! "$(ls -A /mnt/shared/local/${STACK_NAME})" ]; then
-    rsync -aiv /usr/local/vufind/local/ /mnt/shared/local/${STACK_NAME}/local/
-    rsync -aiv /usr/local/vufind/themes/msul/ /mnt/shared/local/${STACK_NAME}/msul/
-    rsync -aiv /usr/local/vufind/module/Catalog/ /mnt/shared/local/${STACK_NAME}/Catalog/
-fi
-rm -rf /usr/local/vufind/local
-ln -sf /mnt/shared/local/${STACK_NAME}/local /usr/local/vufind
-rm -rf /usr/local/vufind/themes/msul
-ln -sf /mnt/shared/local/${STACK_NAME}/msul /usr/local/vufind/themes
-rm -rf /usr/local/vufind/module/Catalog
-ln -sf /mnt/shared/local/${STACK_NAME}/Catalog /usr/local/vufind/module
+if [ "${STACK_NAME}" != "catalog-beta" ]; then
+    mkdir -p /mnt/shared/local/${STACK_NAME}
+    if [ ! "$(ls -A /mnt/shared/local/${STACK_NAME})" ]; then
+        rsync -aiv /usr/local/vufind/local/ /mnt/shared/local/${STACK_NAME}/local/
+        rsync -aiv /usr/local/vufind/themes/msul/ /mnt/shared/local/${STACK_NAME}/msul/
+        rsync -aiv /usr/local/vufind/module/Catalog/ /mnt/shared/local/${STACK_NAME}/Catalog/
+    fi
+    rm -rf /usr/local/vufind/local
+    ln -sf /mnt/shared/local/${STACK_NAME}/local /usr/local/vufind
+    rm -rf /usr/local/vufind/themes/msul
+    ln -sf /mnt/shared/local/${STACK_NAME}/msul /usr/local/vufind/themes
+    rm -rf /usr/local/vufind/module/Catalog
+    ln -sf /mnt/shared/local/${STACK_NAME}/Catalog /usr/local/vufind/module
+fi;
 
 # Ensure SolrCloud is available prior to creating Collections
 SOLR_CLUSTER_SIZE=0
